get all tests running. manifests still fail as expected
