+++ shippo/test/test_shipment.py
@@ -5,12 +5,9 @@ from mock import patch
 
 import shippo
 from shippo.test.helper import (
-    DUMMY_PARCEL,
-    DUMMY_SHIPMENT,
-    FROM_ADDRESS,
+    create_mock_shipment,
     INVALID_SHIPMENT,
     ShippoTestCase,
-    TO_ADDRESS,
 )
 
 
@@ -76,29 +73,6 @@ class ShipmentTests(ShippoTestCase):
     def test_invalid_get_rate(self):
         self.assertRaises(shippo.error.APIError, shippo.Shipment.get_rates,
                           'EXAMPLE_OF_INVALID_ID')
-
-
-def create_mock_shipment():
-    to_address = shippo.Address.create(**TO_ADDRESS)
-    from_address = shippo.Address.create(**FROM_ADDRESS)
-    parcel = shippo.Parcel.create(**DUMMY_PARCEL)
-    SHIPMENT = DUMMY_SHIPMENT.copy()
-    SHIPMENT['address_from'] = from_address.object_id
-    SHIPMENT['address_to'] = to_address.object_id
-    SHIPMENT['parcel'] = parcel.object_id
-    shipment = shippo.Shipment.create(**SHIPMENT)
-    return shipment
-
-
-def create_mock_international_shipment():
-    SHIPMENT = self.create_mock_shipment()
-    customs_item = shippo.CustomsItem.create(**DUMMY_CUSTOMS_ITEM)
-    customs_declaration_parameters = DUMMY_CUSTOMS_DECLARATION.copy()
-    customs_declaration_parameters["items"][0] = customs_item.object_id
-    customs_declaration = shippo.CustomsDeclaration.create(**customs_declaration_parameters)
-    SHIPMENT['customs_declaration'] = customs_declaration.object_id
-    shipment = shippo.Shipment.create(**SHIPMENT)
-    return shipment
 
 
 if __name__ == '__main__':
